Fix test files
