+++ test/unit/core/types/object/test-object.js
@@ -14,36 +14,44 @@
  * limitations under the License.
  */
 
-import * as object from '../../../../src/core/types/object';
+import {
+  deepMerge,
+  getValueForExpr,
+  hasOwn,
+  map,
+  memo,
+  ownProperty,
+  recreateNonProtoObject,
+} from '../../../../../src/core/types/object';
 
 describes.sandboxed('type helpers - objects', {}, () => {
   it('hasOwn', () => {
-    expect(object.hasOwn(object.map(), 'a')).to.be.false;
-    expect(object.hasOwn(object.map({'a': 'b'}), 'b')).to.be.false;
-    expect(object.hasOwn(object.map({'a': {}}), 'a')).to.be.true;
+    expect(hasOwn(map(), 'a')).to.be.false;
+    expect(hasOwn(map({'a': 'b'}), 'b')).to.be.false;
+    expect(hasOwn(map({'a': {}}), 'a')).to.be.true;
   });
 
   it('ownProperty', () => {
-    expect(object.ownProperty({}, '__proto__')).to.be.undefined;
-    expect(object.ownProperty({}, 'constructor')).to.be.undefined;
-    expect(object.ownProperty({foo: 'bar'}, 'foo')).to.equal('bar');
+    expect(ownProperty({}, '__proto__')).to.be.undefined;
+    expect(ownProperty({}, 'constructor')).to.be.undefined;
+    expect(ownProperty({foo: 'bar'}, 'foo')).to.equal('bar');
   });
 
   describe('map', () => {
     it('should make map like objects', () => {
-      expect(object.map().prototype).to.be.undefined;
-      expect(object.map().__proto__).to.be.undefined;
-      expect(object.map().toString).to.be.undefined;
+      expect(map().prototype).to.be.undefined;
+      expect(map().__proto__).to.be.undefined;
+      expect(map().toString).to.be.undefined;
     });
 
     it('should make map like objects from objects', () => {
-      expect(object.map({}).prototype).to.be.undefined;
-      expect(object.map({}).__proto__).to.be.undefined;
-      expect(object.map({}).toString).to.be.undefined;
-      expect(object.map({foo: 'bar'}).foo).to.equal('bar');
+      expect(map({}).prototype).to.be.undefined;
+      expect(map({}).__proto__).to.be.undefined;
+      expect(map({}).toString).to.be.undefined;
+      expect(map({foo: 'bar'}).foo).to.equal('bar');
       const obj = {foo: 'bar', test: 1};
-      expect(object.map(obj).test).to.equal(1);
-      expect(object.map(obj)).to.not.equal(obj);
+      expect(map(obj).test).to.equal(1);
+      expect(map(obj)).to.not.equal(obj);
     });
   });
 
@@ -68,7 +76,7 @@ describes.sandboxed('type helpers - objects', {}, () => {
           },
         },
       };
-      expect(object.deepMerge(destObject, fromObject)).to.deep.equal({
+      expect(deepMerge(destObject, fromObject)).to.deep.equal({
         a: 'hello world',
         b: 'hello world',
         c: {
@@ -93,7 +101,7 @@ describes.sandboxed('type helpers - objects', {}, () => {
           c: ['h', 'i'],
         },
       };
-      expect(object.deepMerge(destObject, fromObject)).to.deep.equal({
+      expect(deepMerge(destObject, fromObject)).to.deep.equal({
         a: [1, 2, 3, 4, 5],
         b: {
           c: ['h', 'i'],
@@ -122,7 +130,7 @@ describes.sandboxed('type helpers - objects', {}, () => {
           },
         },
       };
-      expect(object.deepMerge(destObject, fromObject, 1)).to.deep.equal({
+      expect(deepMerge(destObject, fromObject, 1)).to.deep.equal({
         a: {
           b: {
             c: {
@@ -139,7 +147,7 @@ describes.sandboxed('type helpers - objects', {}, () => {
       destObject.a = destObject;
       const fromObject = {};
       fromObject.a = {};
-      expect(object.deepMerge(destObject, fromObject)).to.deep.equal({
+      expect(deepMerge(destObject, fromObject)).to.deep.equal({
         a: destObject,
       });
     });
@@ -149,7 +157,7 @@ describes.sandboxed('type helpers - objects', {}, () => {
       destObject.a = {};
       const fromObject = {};
       fromObject.a = fromObject;
-      expect(() => object.deepMerge(destObject, fromObject)).to.throw(
+      expect(() => deepMerge(destObject, fromObject)).to.throw(
         /Source object has a circular reference./
       );
     });
@@ -175,7 +183,7 @@ describes.sandboxed('type helpers - objects', {}, () => {
         },
         f: undefined,
       };
-      expect(object.deepMerge(destObject, fromObject)).to.deep.equal({
+      expect(deepMerge(destObject, fromObject)).to.deep.equal({
         a: {
           i: 'j',
         },
@@ -193,7 +201,7 @@ describes.sandboxed('type helpers - objects', {}, () => {
           throw new Error('deep merge tried to merge object with itself');
         },
       };
-      expect(() => object.deepMerge(destObject, destObject)).to.not.throw();
+      expect(() => deepMerge(destObject, destObject)).to.not.throw();
     });
   });
 
@@ -217,15 +225,127 @@ describes.sandboxed('type helpers - objects', {}, () => {
 
     it('should allocate object on first and first use only', () => {
       // First access: allocate and reuse.
-      expect(object.memo(obj, PROP, factory)).to.equal('OBJ:1');
-      expect(object.memo(obj, PROP, factory)).to.equal('OBJ:1');
+      expect(memo(obj, PROP, factory)).to.equal('OBJ:1');
+      expect(memo(obj, PROP, factory)).to.equal('OBJ:1');
 
       // Same object, different property: allocate again.
-      expect(object.memo(obj, PROP + '2', factory)).to.equal('OBJ:2');
-      expect(object.memo(obj, PROP + '2', factory)).to.equal('OBJ:2');
+      expect(memo(obj, PROP + '2', factory)).to.equal('OBJ:2');
+      expect(memo(obj, PROP + '2', factory)).to.equal('OBJ:2');
 
       // A new object: allocate again.
-      expect(object.memo({name: 'OBJ'}, PROP, factory)).to.equal('OBJ:3');
+      expect(memo({name: 'OBJ'}, PROP, factory)).to.equal('OBJ:3');
+    });
+  });
+
+  describe('getValueForExpr', () => {
+    it('should return self for "."', () => {
+      const obj = {str: 'A', num: 1, bool: true, val: null};
+      expect(getValueForExpr(obj, '.')).to.equal(obj);
+    });
+
+    it('should return a simple value', () => {
+      const obj = {str: 'A', num: 1, bool: true, val: null};
+      expect(getValueForExpr(obj, 'str')).to.equal('A');
+      expect(getValueForExpr(obj, 'num')).to.equal(1);
+      expect(getValueForExpr(obj, 'bool')).to.equal(true);
+      expect(getValueForExpr(obj, 'val')).to.be.null;
+      expect(getValueForExpr(obj, 'other')).to.be.undefined;
+    });
+
+    it('should return a nested value', () => {
+      const child = {str: 'A', num: 1, bool: true, val: null};
+      const obj = {child};
+      expect(getValueForExpr(obj, 'child')).to.deep.equal(child);
+      expect(getValueForExpr(obj, 'child.str')).to.equal('A');
+      expect(getValueForExpr(obj, 'child.num')).to.equal(1);
+      expect(getValueForExpr(obj, 'child.bool')).to.equal(true);
+      expect(getValueForExpr(obj, 'child.val')).to.be.null;
+      expect(getValueForExpr(obj, 'child.other')).to.be.undefined;
+    });
+
+    it('should return a nested value without proto', () => {
+      const child = {str: 'A', num: 1, bool: true, val: null};
+      const obj = recreateNonProtoObject({child});
+      expect(getValueForExpr(obj, 'child')).to.deep.equal(child);
+      expect(getValueForExpr(obj, 'child.str')).to.equal('A');
+      expect(getValueForExpr(obj, 'child.num')).to.equal(1);
+      expect(getValueForExpr(obj, 'child.bool')).to.equal(true);
+      expect(getValueForExpr(obj, 'child.val')).to.be.null;
+      expect(getValueForExpr(obj, 'child.other')).to.be.undefined;
+    });
+
+    it('should shortcircuit if a parent in chain missing', () => {
+      const child = {str: 'A'};
+      const obj = {child};
+      expect(getValueForExpr(obj, 'child.str')).to.equal('A');
+      expect(getValueForExpr(obj, 'unknown.str')).to.be.undefined;
+      expect(getValueForExpr(obj, 'unknown.chain.str')).to.be.undefined;
+    });
+
+    it('should shortcircuit if a parent in chain is not an object', () => {
+      const child = {str: 'A'};
+      const obj = {child, nonobj: 'B'};
+      expect(getValueForExpr(obj, 'child.str')).to.equal('A');
+      expect(getValueForExpr(obj, 'nonobj')).to.equal('B');
+      expect(getValueForExpr(obj, 'nonobj.str')).to.be.undefined;
+    });
+
+    it('should only search in own properties', () => {
+      const ancestor = {num: 1};
+      const obj = Object.create(ancestor);
+      obj.str = 'A';
+      expect(getValueForExpr(obj, 'str')).to.equal('A');
+      expect(getValueForExpr(ancestor, 'num')).to.equal(1);
+      expect(getValueForExpr(obj, 'num')).to.be.undefined;
+      expect(getValueForExpr(obj, '__proto__')).to.be.undefined;
+    });
+
+    it('should support array index', () => {
+      const child = {num: 1, str: 'A'};
+      const obj = {foo: [child]};
+      expect(getValueForExpr(obj, 'foo.0.num')).to.equal(1);
+      expect(getValueForExpr(obj, 'foo.0.str')).to.equal('A');
+      expect(getValueForExpr(obj, 'foo.0a.str')).to.be.undefined;
+      expect(getValueForExpr(obj, 'foo.1.num')).to.be.undefined;
+      expect(getValueForExpr(obj, 'foo.1.str')).to.be.undefined;
+    });
+
+    it('should only search in own properties of arrays', () => {
+      const arr = ['A'];
+      expect(getValueForExpr(arr, '0')).to.equal('A');
+      expect(getValueForExpr(arr, '1')).to.be.undefined;
+      expect(getValueForExpr(arr, 'concat')).to.be.undefined;
+      expect(getValueForExpr(arr, '__proto__')).to.be.undefined;
+    });
+  });
+
+  describe('recreateNonProtoObject', () => {
+    it('should recreate an empty object', () => {
+      const original = {};
+      const copy = recreateNonProtoObject(original);
+      expect(copy).to.deep.equal(original);
+      expect(copy).to.not.equal(original);
+      expect(copy.__proto__).to.be.undefined;
+    });
+
+    it('should recreate an object', () => {
+      const original = {str: 'A', num: 1, bool: true, val: null};
+      const copy = recreateNonProtoObject(original);
+      expect(copy).to.deep.equal(original);
+      expect(copy).to.not.equal(original);
+      expect(copy.__proto__).to.be.undefined;
+      expect(copy.val).to.be.null;
+    });
+
+    it('should recreate a nested object', () => {
+      const original = {child: {str: 'A', num: 1, bool: true, val: null}};
+      const copy = recreateNonProtoObject(original);
+      expect(copy).to.deep.equal(original);
+      expect(copy).to.not.equal(original);
+      expect(copy.__proto__).to.be.undefined;
+      expect(copy.child).to.deep.equal(original.child);
+      expect(copy.child).to.not.equal(original.child);
+      expect(copy.child.__proto__).to.be.undefined;
     });
   });
 });
